ngfw-15794:Relocate logo-upload extension test to branding_manager and fix WireGuard tunnel network conflict
